+++ components/Wizard/Steps/MainStep.tsx
@@ -3,39 +3,27 @@ import { ImmutableXClient } from "@imtbl/imx-sdk";
 import { useWeb3React } from "@web3-react/core";
 import { Field, Form, Formik, FormikErrors, FormikProps, useField, useFormikContext } from "formik";
 import { FC, useCallback, useEffect, useRef, useState } from "react";
-import { useAccountState } from "../../../context/account";
 import { useQueryState } from "../../../context/query";
 import { useSettingsState } from "../../../context/settings";
 import { CryptoNetwork } from "../../../Models/CryptoNetwork";
 import { Currency } from "../../../Models/Currency";
 import { Exchange } from "../../../Models/Exchange";
-import { Partner } from "../../../Models/Partner";
 import { SwapFormValues } from "../../DTOs/SwapFormValues";
-import InsetSelectMenu from "../../selectMenu/insetSelectMenu";
-import SelectMenu from "../../selectMenu/selectMenu";
 import { SelectMenuItem } from "../../selectMenu/selectMenuItem";
 import Image from 'next/image'
 import SwapButton from "../../buttons/swapButton";
-import { useWizardState } from "../../../context/wizard";
 import { useSwapDataUpdate } from "../../../context/swap";
 import Select from "../../Select/Select";
 import React from "react";
-import { useInterval } from "../../../hooks/useInyterval";
-import { useFormWizardaUpdate, useFormWizardState } from "../../../context/formWizardProvider";
+import { useFormWizardaUpdate } from "../../../context/formWizardProvider";
 import { ExchangeAuthorizationSteps, FormWizardSteps } from "../../../Models/Wizard";
 import TokenService from "../../../lib/TokenService";
 import { useUserExchangeDataUpdate } from "../../../context/userExchange";
 import axios from "axios";
-import LayerSwapAuthApiClient from "../../../lib/userAuthApiClient";
-import { ExclamationIcon } from "@heroicons/react/outline";
-import AmountAndFeeDetails from "../../amountAndFeeDetailsComponent";
+import AmountAndFeeDetails from "../../Disclosure/amountAndFeeDetailsComponent";
 import ConnectImmutableX from "./ConnectImmutableX";
 import ConnectDeversifi from "../../ConnectDeversifi";
-import SendFeedback from "../../sendFeedback";
-import SlideOver, { SildeOverRef } from "../../SlideOver";
-import { DocIframe } from "../../docInIframe";
 import toast from "react-hot-toast";
-import { BlacklistedAddress } from "../../../Models/BlacklistedAddress";
 import { InjectedConnector } from "@web3-react/injected-connector";
 import { isValidAddress } from "../../../lib/addressValidator";
 import { clearTempData, getTempData } from "../../../lib/openLink";
